moved firebase into firebase package
